+++ src/Faithlife.Build/DotNetBuildSettings.cs
@@ -35,25 +35,11 @@ namespace Faithlife.Build
 		public DotNetDocsSettings DocsSettings { get; set; }
 
 		/// <summary>
-		/// The version of the <c>sourcelink</c> tool to use when testing packages.
-		/// </summary>
-		/// <remarks>Defaults to a stable version, which may change with new versions of <b>Faithlife.Build</b>,
-		/// but will not change unless <b>Faithlife.Build</b> is updated.</remarks>
-		public string SourceLinkToolVersion { get; set; }
-
-		/// <summary>
 		/// The options and flags used by <see cref="DotNetBuild"/>.
 		/// </summary>
 		/// <remarks>Any properties not set before <see cref="DotNetBuild.AddDotNetTargets"/> is called
 		/// will be set afterward.</remarks>
 		public DotNetBuildOptions BuildOptions { get; set; }
-
-		/// <summary>
-		/// A function that returns true if the named project uses SourceLink.
-		/// </summary>
-		/// <remarks>If not specified, all projects are assumed to use SourceLink. This property
-		/// determines whether the <c>sourcelink</c> tool is used to test each package.</remarks>
-		public Func<string, bool> ProjectUsesSourceLink { get; set; }
 
 		/// <summary>
 		/// Set to use <c>MSBuild</c> instead of <c>dotnet</c> to build the solution.
@@ -71,20 +57,6 @@ namespace Faithlife.Build
 		public string SolutionPlatform { get; set; }
 
 		/// <summary>
-		/// A function that returns true if the named project uses semantic versioning.
-		/// </summary>
-		/// <remarks>If not specified, all projects are assumed to use semantic versioning. This property
-		/// determines whether the <c>packagediff</c> tool is used to test each package.</remarks>
-		public Func<string, bool> ProjectUsesSemVer { get; set; }
-
-		/// <summary>
-		/// The version of the <c>packagediff</c> tool to use when testing packages.
-		/// </summary>
-		/// <remarks>Defaults to a stable version, which may change with new versions of <b>Faithlife.Build</b>,
-		/// but will not change unless <b>Faithlife.Build</b> is updated.</remarks>
-		public string PackageDiffToolVersion { get; set; }
-
-		/// <summary>
 		/// The maximum number of CPUs to use when building.
 		/// </summary>
 		/// <remarks>Use <c>1</c> to enforce sequential builds.</remarks>
@@ -98,6 +70,11 @@ namespace Faithlife.Build
 		public Func<string, IEnumerable<(string Key, string Value)>> ExtraProperties { get; set; }
 
 		/// <summary>
+		/// The SourceLink settings. Must be set to test SourceLink URLs.
+		/// </summary>
+		public SourceLinkSettings SourceLinkSettings { get; set; }
+
+		/// <summary>
 		/// Clones the settings.
 		/// </summary>
 		public DotNetBuildSettings Clone()
@@ -106,6 +83,7 @@ namespace Faithlife.Build
 			clone.DocsSettings = clone.DocsSettings?.Clone();
 			clone.MSBuildSettings = clone.MSBuildSettings?.Clone();
 			clone.TestSettings = clone.TestSettings?.Clone();
+			clone.SourceLinkSettings = clone.SourceLinkSettings?.Clone();
 			return clone;
 		}
 	}