+++ VersionHistory.md
@@ -8,6 +8,12 @@ Prefix the description of the change with `[major]`, `[minor]`, or `[patch]` in 
 
 ## Released
 
+### 3.0.0
+
+* [major] Drop support for `packagediff` for now. (Too buggy.)
+* [patch] `package` target depends on `clean` for simpler package detection.
+* [major] Support SourceLink authentication, moving settings into `SourceLinkSettings`. Now disabled by default.
+
 ### 2.2.0
 
 * [minor] Add `DotNetBuildProperties.ExtraProperties`.